+++ Editor/Unity.Platforms/BuildTarget.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
+using Newtonsoft.Json.Linq;
 
 namespace Unity.Platforms
 {
@@ -112,6 +113,8 @@ namespace Unity.Platforms
         {
             throw new NotImplementedException();
         }
+
+        public virtual void WriteBeeConfigFile(string path) {}
     }
 
     internal sealed class UnknownBuildTarget : BuildTarget
@@ -164,5 +167,10 @@ namespace Unity.Platforms
                 return exception.Types.Where(type => type != null);
             }
         }
+    }
+
+    public interface IDotsRuntimeBuildModifier
+    {
+        void Modify(JObject settingsJObject);
     }
 }